Refactors the parser, adds more tests, removes SECRET_KEY env for tests
The parser has been refactored to be more maintainable and a change to the regex was made to fix the error in issue #39. Tests have been written to cover this (and a few other) edge cases that weren't previously covered.

The need for SECRET_KEY used by the migrations has been removed for any of the testing scripts.